requirements.txt
GitPython==3.1.30
flake8==5.0.4
black==23.3.0
openai==0.27.8 